requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
astroid==3.0.2
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
coverage==7.4.1
dataclasses-json==0.6.4
dill==0.3.8
distro==1.9.0
exceptiongroup==1.2.0
Faker==23.1.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
iniconfig==2.0.0
install==1.3.5
isort==5.13.2
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.9
langchain-community==0.0.24
langchain-core==0.1.26
langchain-openai==0.0.7
langsmith==0.1.6
marshmallow==3.20.2
mccabe==0.7.0
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.12.0
orjson==3.9.15
packaging==23.2
platformdirs==4.1.0
pluggy==1.4.0
pydantic==2.6.2
pydantic_core==2.16.3
pylint==3.0.3
pylint-pytest==1.1.7
pytest==8.0.0
pytest-cov==4.1.0
pytest-pylint==0.21.0
python-dateutil==2.8.2
python-dotenv==1.0.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.27
tenacity==8.2.3
tiktoken==0.6.0
tomli==2.0.1
tomlkit==0.12.3
tqdm==4.66.2
typing-inspect==0.9.0
typing_extensions==4.9.0
urllib3==2.2.1
yarl==1.9.4
